feat(runtime-c-api) Extract the `memory` module.
